refactor inline logic into helper functions
